backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
kafka-python==2.2.15
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
minio==7.2.16
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.37.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
